cmd/ansible-operator,cmd/helm-operator: embed code in main.go (#3415)
